create new types and filter hook
create types for issue filters that will be passed to query and enums for the status of issues.
create filter hook for central local of resources used in issue-filter and issue-list.  The hook stores the filter results of the user as well as the function that handles the change in the selection of filter