Issue list filters (#23)
* fixed wrong font size of button

* refactored Select component

using Label/Input tags was becoming a bit cumbersome. I replaced label/input structure with spans/divs
I also refactored 'click away'. I replaced it with the 'onBlur' rather than hard coding

* refactored Input element

As it stood, label and input tags as siblings were causing UI issues because of flex display. When I nested the input tag inside label, the issue went away.
I also fixed some errors on css properties concerning font and color, etc.

* initial simple structure for issue filters

I created the button on the left and the filters on the right.

* moved width to label and removed unused wrapper

* fixed UI issue

fixed onBlur issue by adding tabIndex attribute
adjusted width to allow for more control given to parent

* fixed layout issues with select components showing above fixed component

Select component had relative position and thus conflicted with fixed position of navigation.  Added z-index.

* fixed layout for filter container

changed order of rendering when in desktop mode.  default mode is mobile.
allowed for changing width of components

* removed log and uncommented onBlur event

* adjust syntax and add hook

added a filter hook so the application could use the code in both issue-filter and issue-list

* create new types and filter hook

create types for issue filters that will be passed to query and enums for the status of issues.
create filter hook for central local of resources used in issue-filter and issue-list.  The hook stores the filter results of the user as well as the function that handles the change in the selection of filter

* incorporate filter hook and filter object

filters chosen by user are added to the query params for next page navigation and for initial getting of issues.

* update useGetIssue hook and api call for getting issues

added filters to params in the query for getting issues and in the database call.

* fix sytax and refactor selection verification

isSelected previously passed in string, making it unreadable. I was comparing value with label rather than value with value. Refactored by passing in object and made clear we are comparing value with value.

* made refactoring and added search capability

I refactoed error and status options by placing them in a utils folder
I added a function to handle input change by user.  It incorporates a debounce method that waits 500ms or 1 second before processing the user's input

I also had to add 'undefined' value to accomodate unselected option.

* fixed sytax and refactored unordered list key value

* added checkbox and refactored checkbox component

I made children of Checkbox optional.  Input element was being squished with flex shrink so I added a flex display to input element and set it to not shrink.

* updated issue-list to include desktop chart

* remove checkbox and fix tag

* added proper url paramters